Moves assert_compare.h file to test folder.
