+++ test/api/compare_test.cc
@@ -33,8 +33,8 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/math/overlapping_type.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
