+++ test/api/lane_data_test.cc
@@ -31,6 +31,7 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/common/assertion_error.h"
 #include "maliput/common/compare.h"
@@ -39,7 +40,6 @@
 #include "maliput/math/quaternion.h"
 #include "maliput/math/vector.h"
 #include "maliput/test_utilities/mock.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
