+++ test/api/phase_test.cc
@@ -36,9 +36,9 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/rules/compare.h"
 #include "maliput/common/assertion_error.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
