+++ test/api/range_value_rule_state_provider_test.cc
@@ -31,11 +31,11 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/rules/compare.h"
 #include "maliput/api/rules/range_value_rule.h"
 #include "maliput/api/rules/rule.h"
 #include "maliput/test_utilities/mock.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
