+++ test/api/regions_test.cc
@@ -34,11 +34,11 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/common/assertion_error.h"
 #include "maliput/common/maliput_throw.h"
 #include "maliput/test_utilities/mock.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
