+++ test/api/right_of_way_rule_state_provider_test.cc
@@ -34,8 +34,8 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/rules/compare.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
