+++ test/api/rules_direction_usage_test.cc
@@ -37,10 +37,10 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/api/regions.h"
 #include "maliput/api/rules/compare.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
