+++ test/api/rules_right_of_way_test.cc
@@ -34,12 +34,12 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/api/regions.h"
 #include "maliput/api/rules/compare.h"
 #include "maliput/common/assertion_error.h"
 #include "maliput/test_utilities/mock.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
