+++ test/api/rules_speed_limit_test.cc
@@ -34,11 +34,11 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/api/regions.h"
 #include "maliput/api/rules/compare.h"
 #include "maliput/common/assertion_error.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace api {
