+++ test/base/manual_rulebook_test.cc
@@ -31,13 +31,13 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/regions.h"
 #include "maliput/api/rules/compare.h"
 #include "maliput/api/rules/right_of_way_rule.h"
 #include "maliput/api/rules/speed_limit_rule.h"
 #include "maliput/common/assertion_error.h"
 #include "maliput/test_utilities/mock.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace test {
