+++ test/base/traffic_light_book_test.cc
@@ -31,9 +31,9 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/lane_data.h"
 #include "maliput/api/rules/compare.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace {
