+++ test/geometry_base/filter_positions_test.cc
@@ -33,10 +33,10 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/api/lane.h"
 #include "maliput/api/lane_data.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace geometry_base {
