+++ test/geometry_base/geometry_base_test.cc
@@ -35,9 +35,9 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/common/maliput_unused.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace geometry_base {
