+++ test/math/matrix_test.cc
@@ -31,8 +31,8 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/math/compare.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace math {
