+++ test/math/quaternion_test.cc
@@ -215,12 +215,12 @@
 
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/common/compare.h"
 #include "maliput/math/compare.h"
 #include "maliput/math/matrix.h"
 #include "maliput/math/roll_pitch_yaw.h"
 #include "maliput/math/vector.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace math {
