+++ test/routing/phase_test.cc
@@ -34,13 +34,13 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
+#include "assert_compare.h"
 #include "maliput/api/compare.h"
 #include "maliput/api/lane_data.h"
 #include "maliput/api/regions.h"
 #include "maliput/common/assertion_error.h"
 #include "maliput/routing/compare.h"
 #include "routing/road_network_mocks.h"
-#include "test_utilities/assert_compare.h"
 
 namespace maliput {
 namespace routing {
